fix client url config
when the client is accessed through a browser on the host, the
javascript on the page is running on the host, and therefore does not
actually have access to the `front` network on the name-resolution it
provides. The client must be configured to use the endpoints exposed
on localhost in this case.